Adding playwright
